+++ src/actions/category/EditAction.php
@@ -17,10 +17,10 @@ namespace blackcube\admin\actions\category;
 use blackcube\admin\actions\BaseElementAction;
 use blackcube\admin\helpers\Category as CategoryHelper;
 use blackcube\admin\models\SlugForm;
-use blackcube\core\models\Category;
+use blackcube\admin\Module;
+use blackcube\core\interfaces\PluginHookInterface;
+use blackcube\core\interfaces\PluginsHandlerInterface;
 use blackcube\core\models\Language;
-use blackcube\core\models\Type;
-use yii\base\Action;
 use yii\web\NotFoundHttpException;
 use yii\web\Response;
 use Yii;
@@ -62,19 +62,37 @@ class EditAction extends BaseElementAction
         if ($category === null) {
             throw new NotFoundHttpException();
         }
+        $pluginsHandler = Yii::createObject(PluginsHandlerInterface::class);
+        /* @var $pluginsHandler \blackcube\core\interfaces\PluginsHandlerInterface */
+        $pluginsHandler->runHook(PluginHookInterface::PLUGIN_HOOK_LOAD, $category);
+
         $slugForm = Yii::createObject([
             'class' => SlugForm::class,
             'element' => $category
         ]);
         $blocs = $category->getBlocs()->all();
+        $transaction = Module::getInstance()->db->beginTransaction();
         $result = CategoryHelper::saveElement($category, $blocs, $slugForm);
-        if ($result === true) {
-            return $this->controller->redirect([$this->targetAction, 'id' => $category->id]);
+        $validatePlugins = $pluginsHandler->runHook(PluginHookInterface::PLUGIN_HOOK_VALIDATE, $category);
+        $validatePlugins = array_reduce($validatePlugins, function($accumulator, $item) {
+            return $accumulator && $item;
+        }, true);
+        if ($result === true && $validatePlugins === true) {
+            $savePlugins = $pluginsHandler->runHook(PluginHookInterface::PLUGIN_HOOK_SAVE, $category);
+            $savePlugins = array_reduce($savePlugins, function($accumulator, $item) {
+                return $accumulator && $item;
+            }, true);
+            if ($savePlugins === true) {
+                $transaction->commit();
+                return $this->controller->redirect([$this->targetAction, 'id' => $category->id]);
+            }
         }
+        $transaction->rollBack();
         $languagesQuery = Language::find()->active()->orderBy(['name' => SORT_ASC]);
         $typesQuery = $this->getTypesQuery()
             ->orderBy(['name' => SORT_ASC]);
         return $this->controller->render($this->view, [
+            'pluginsHandler' => $pluginsHandler,
             'category' => $category,
             'slugForm' => $slugForm,
             'typesQuery' => $typesQuery,