+++ src/controllers/MenuController.php
@@ -26,24 +26,11 @@ use blackcube\admin\actions\menu\UpItemAction;
 use blackcube\admin\actions\ToggleAction;
 use blackcube\admin\actions\ModalAction;
 use blackcube\admin\components\Rbac;
-use blackcube\admin\Module;
-use blackcube\core\models\Category;
-use blackcube\core\models\Composite;
-use blackcube\core\models\Node;
-use blackcube\core\models\Tag;
-use blackcube\core\Module as CoreModule;
-use blackcube\core\interfaces\BlackcubeControllerInterface;
-use blackcube\core\models\Language;
 use blackcube\core\models\Menu;
 use blackcube\core\models\MenuItem;
-use yii\base\ErrorException;
-use yii\data\ActiveDataProvider;
 use yii\filters\AccessControl;
 use yii\filters\AjaxFilter;
-use yii\helpers\Inflector;
 use yii\web\Controller;
-use yii\web\NotFoundHttpException;
-use yii\web\Response;
 use Yii;
 
 /**