+++ src/controllers/PluginController.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * SlugController.php
+ * PluginController.php
  *
  * PHP version 7.2+
  *
@@ -15,18 +15,20 @@
 namespace blackcube\admin\controllers;
 
 use blackcube\admin\actions\ModalAction;
-use blackcube\admin\actions\slug\DeleteAction;
-use blackcube\admin\actions\slug\EditAction;
-use blackcube\admin\actions\slug\IndexAction;
-use blackcube\admin\actions\slug\ToggleAction;
+use blackcube\admin\actions\plugin\IndexAction;
+use blackcube\admin\actions\plugin\ToggleAction;
+use blackcube\admin\actions\plugin\ToggleRegisterAction;
 use blackcube\admin\components\Rbac;
-use blackcube\core\models\Slug;
+use blackcube\core\interfaces\PluginManagerConfigurableInterface;
+use blackcube\core\interfaces\PluginsHandlerInterface;
+use blackcube\core\models\Plugin;
 use yii\filters\AccessControl;
 use yii\filters\AjaxFilter;
+use yii\web\Controller;
 use Yii;
 
 /**
- * Class SlugController
+ * Class PluginController
  *
  * @author Philippe Gaultier <pgaultier@redcat.io>
  * @copyright 2010-2020 Redcat
@@ -35,7 +37,7 @@ use Yii;
  * @link https://www.redcat.io
  * @package blackcube\admin\controllers
  */
-class SlugController extends BaseElementController
+class PluginController extends Controller
 {
     /**
      * {@inheritDoc}
@@ -51,28 +53,20 @@ class SlugController extends BaseElementController
                     'actions' => [
                         'modal', 'index',
                     ],
-                    'roles' => [Rbac::PERMISSION_SLUG_VIEW],
+                    'roles' => [Rbac::PERMISSION_PLUGIN_VIEW],
                 ],
                 [
                     'allow' => true,
                     'actions' => [
-                        'toggle', 'edit',
+                        'edit', 'toggle', 'toggle-register',
                     ],
-                    'roles' => [Rbac::PERMISSION_SLUG_UPDATE],
+                    'roles' => [Rbac::PERMISSION_PLUGIN_UPDATE],
                 ],
-                [
-                    'allow' => true,
-                    'actions' => [
-                        'delete',
-                    ],
-                    'roles' => [Rbac::PERMISSION_SLUG_DELETE],
-                ],
-
             ]
         ];
         $behaviors['forceAjax'] = [
             'class' => AjaxFilter::class,
-            'only' => ['modal', 'toggle'],
+            'only' => ['modal', 'actions'],
         ];
         return $behaviors;
     }
@@ -83,24 +77,31 @@ class SlugController extends BaseElementController
     public function actions()
     {
         $actions = parent::actions();
+        $actions['toggle'] = [
+            'class' => ToggleAction::class
+        ];
+        $actions['toggle-register'] = [
+            'class' => ToggleRegisterAction::class
+        ];
         $actions['modal'] = [
             'class' => ModalAction::class,
-            'elementClass' => Slug::class
+            'elementClass' => Plugin::class
         ];
         $actions['index'] = [
             'class' => IndexAction::class,
         ];
-        $actions['edit'] = [
-            'class' => EditAction::class,
-        ];
-        $actions['delete'] = [
-            'class' => DeleteAction::class,
-        ];
-        $actions['toggle'] = [
-            'class' => ToggleAction::class,
-        ];
+        $id = Yii::$app->request->get('id', null);
+        if ($id !== null) {
+            $pluginsHandler = Yii::createObject(PluginsHandlerInterface::class);
+            $pluginManager = $pluginsHandler->getPluginManager($id);
+            if ($pluginManager !== null && $pluginManager instanceof PluginManagerConfigurableInterface) {
+                $configureAction = $pluginManager->getConfigureAction();
+                if ($configureAction !== null) {
+                    $actions['edit'] = $configureAction;
+                }
+            }
+            /**/
+        }
         return $actions;
     }
-
-
 }