+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Blackcube Admin - CHANGELOG 
 ===========================
 
+ * Add plugins management
+ * Enable plugins in `node`, `composite`, `category` and `tag`
+ * Enable plugins widgets in `node`, `composite`, `category` and `tag`
+ * Update Rbac roles and permissions
+ * Update locale according to new code
+ 
 Release 1.1.0 June, 10th 2020 <pgaultier@redcat.io>
 ---------------------------------------------------